升级 knowledge-net-store 版本
