+++ config/initializers/gem_integration.rb
@@ -149,17 +149,6 @@ module KnowledgeNetPlanStore
   end
 end
 
-module KnowledgeNetStore
-  class Point
-    include PinyinSearch
-
-    has_and_belongs_to_many :tutorials,
-                            :class_name => "KnowledgeNetPlanStore::Tutorial"
-
-    pinyin :name
-  end
-end
-
 # -----------------------------------------
 # virtual_file_system 相关
 module FileEntityVFSModule
@@ -367,47 +356,7 @@ module DocumentsStore
   end
 end
 
-module KnowledgeNetStore
-  class Net
-    has_many :documents,
-             :class_name => 'DocumentsStore::Document',
-             :dependent => :destroy
-
-    has_many :plans,
-             :class_name => 'KnowledgeNetPlanStore::Plan',
-             :dependent => :destroy
-
-    has_many :virtual_files,
-             :class_name => 'VirtualFileSystem::File',
-             :dependent => :destroy
-
-    after_create :create_default_plan
-
-    def default_plan
-      create_default_plan if plans.blank?
-      plans.first
-    end
-
-    def topics
-      default_plan.topics
-    end
-
-    private
-
-    def create_default_plan
-      self.plans.create :title => "default_plan"
-    end
-  end
-
-  class Point
-    has_and_belongs_to_many :virtual_files,
-                            :class_name => 'VirtualFileSystem::File',
-                            :inverse_of => :points
-  end
-end
-
 [
-  KnowledgeNetStore::Net,
   VirtualFileSystem::File,
   KnowledgeNetPlanStore::Topic,
   KnowledgeNetPlanStore::Tutorial
