+++ config/initializers/model_names.rb
@@ -1,17 +1,3 @@
-module KnowledgeNetStore
-  class Net
-    def self.model_name
-      ActiveModel::Name.new(KnowledgeNetStore::Net, nil, 'net')
-    end
-  end
-
-  class Point
-    def self.model_name
-      ActiveModel::Name.new(KnowledgeNetStore::Point, nil, 'point')
-    end
-  end
-end
-
 module KnowledgeNetPlanStore
   class Plan
     def self.model_name
